refactor functions back to kubeCrpyt
