+++ lib/kubeCrypt.ts
@@ -14,14 +14,17 @@
  * limitations under the License.
  */
 
+import { guid } from "@atomist/automation-client";
 import * as k8s from "@kubernetes/client-node";
+import * as fs from "fs-extra";
+import inquirer = require("inquirer");
+import * as yaml from "js-yaml";
 import * as _ from "lodash";
 import { DeepPartial } from "ts-essentials";
+import { maskString } from "./config";
 import {
     base64,
     crypt,
-    handleSecretKeyParameter,
-    handleSecretParameter,
     printSecret,
 } from "./kubeUtils";
 import * as print from "./print";
@@ -77,4 +80,65 @@ export async function kubeCrypt(opts: KubeCryptOptions): Promise<number> {
     }
 
     return 0;
+}
+
+/**
+ * Handle the literal or file secret parameter from the cli
+ * @param opts file or literal of KubeCryptOptions
+ * @throws error if the yaml cannot be loaded
+ * @returns secret
+ */
+export async function handleSecretParameter(opts: Pick<KubeCryptOptions, "file" | "literal" | "action">): Promise<DeepPartial<k8s.V1Secret>> {
+    let secret: DeepPartial<k8s.V1Secret>;
+    const literalProp = `literal-${guid()}`;
+    if (opts.literal) {
+        secret = wrapLiteral(opts.literal, literalProp);
+    } else if (opts.file) {
+        const secretString = await fs.readFile(opts.file, "utf8");
+        secret = await yaml.safeLoad(secretString);
+    } else {
+        const answers = await inquirer.prompt<Record<string, string>>([{
+            type: "input",
+            name: "literal",
+            message: `Enter literal string to be ${opts.action}ed:`,
+        }]);
+        secret = wrapLiteral(answers.literal, literalProp);
+    }
+    return secret;
+}
+
+/**
+ *  Creates a k8s.V1Secret with the input in the data section.
+ * @param literal String to wrap in k8s.V1Secret
+ * @param prop property name
+ * @returns the k8s.V1Secret
+ */
+function wrapLiteral(literal: string, prop: string): DeepPartial<k8s.V1Secret> {
+    const secret: DeepPartial<k8s.V1Secret> = {
+        apiVersion: "v1",
+        data: {},
+        kind: "Secret",
+        type: "Opaque",
+    };
+    secret.data[prop] = literal;
+    return secret;
+}
+
+/**
+ * Handle the secret key parameter from the cli
+ * @param opts secretKey from KubeCryptOptions
+ * @returns the secret
+ */
+export async function handleSecretKeyParameter(opts: Pick<KubeCryptOptions, "secretKey">): Promise<string> {
+    if (!opts.secretKey) {
+        const answers = await inquirer.prompt<Record<string, string>>([{
+            type: "input",
+            name: "secretKey",
+            message: `Enter encryption key:`,
+            transformer: maskString,
+            validate: v => v.length < 1 ? "Secret key must have non-zero length" : true,
+        }]);
+        opts.secretKey = answers.secretKey;
+    }
+    return opts.secretKey;
 }
